requirements.txt
fastapi
pydantic
uvicorn
mangum  
boto3
uuid
python-multipart
requests
python-dotenv
openai
pytest
httpx
awsebcli